feat: IVP's rotation in doubles, pinned to vphysics.dll's own routines called in process (B369, D172)
IvpQuaternion.Product, Normalise, Interpolate, Delta and SineDelta and IvpIntegrator.Rotate port FUN_180070d60, FUN_180070c60, FUN_180071060, FUN_180071680, FUN_180070f50 and FUN_180099fc0 instruction for instruction, every destination named. A core's two orientations are doubles, as core+0x180 and +0x1a0 are.

The seven divergences B369 filed are gone: the fixed-point normalization, the float series grouping with no clamp, the new delta on the left of each later sub-step, the float spin sum and double Euler products, the dot and renormalization grouping with vphysics' own acos and sin, and the second route for bit 0x8, phase 5 and the core+0x58 axis.

The vphysics-rotation probe compares 50,000 cases on both sin paths, none differing; IvpRotationConformanceTests replays 400 and 102 NaN pairs. The old off-unit Normalise case used (0, 0, 0, 4), on which the binary's loop never returns.

+++ managed/Tf2DemoSalvage.Animation/Animating/IvpMatrix.cs
@@ -41,12 +41,11 @@ public readonly record struct IvpMatrix(
     /// m8 = x·2z − w·2y         m9 = w·2x + y·2z         m10 = 1 − (y·2y + x·2x)
     /// </code>
     ///
-    /// **Every term in doubles**, from a float rotation widened first, as the engine's rotation at
-    /// `core+0x180` is stored in doubles. The indices skip 3, 7 and 11 because the engine's storage is a
-    /// 4×4 and those are its unused fourth column.
+    /// **Every term in doubles**, as the engine's rotation at `core+0x180` is stored in doubles. The indices skip 3, 7 and 11
+    /// because the engine's storage is a 4×4 and those are its unused fourth column.
     /// </remarks>
     public static IvpMatrix FromRotation(
-        (float X, float Y, float Z, float W) rotation, (double X, double Y, double Z) position)
+        (double X, double Y, double Z, double W) rotation, (double X, double Y, double Z) position)
     {
         double x = rotation.X;
         double y = rotation.Y;
